Move templatetag to core
